+++ service/worker/domainReplicationTaskHandler_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/uber/cadence/.gen/go/replicator"
 	"github.com/uber/cadence/.gen/go/shared"
 	"github.com/uber/cadence/common"
-	"github.com/uber/cadence/common/mocks"
 	"github.com/uber/cadence/common/persistence"
 )
 
@@ -41,7 +40,6 @@ type (
 		suite.Suite
 		persistence.TestBase
 		domainReplicator *domainReplicatorImpl
-		kafkaProducer    *mocks.KafkaProducer
 	}
 )
 
@@ -66,147 +64,10 @@ func (s *domainReplicatorSuite) SetupTest() {
 		s.MetadataManager,
 		bark.NewLoggerFromLogrus(logrus.New()),
 	).(*domainReplicatorImpl)
-	s.kafkaProducer = &mocks.KafkaProducer{}
 }
 
 func (s *domainReplicatorSuite) TearDownTest() {
 	s.TearDownWorkflowStore()
-}
-
-func (s *domainReplicatorSuite) TestHandleTransmissionTask_RegisterDomainTask() {
-	taskType := replicator.ReplicationTaskTypeDomain
-	id := uuid.New()
-	name := "some random domain test name"
-	status := shared.DomainStatusRegistered
-	description := "some random test description"
-	ownerEmail := "some random test owner"
-	retention := int32(10)
-	emitMetric := true
-	clusterActive := "some random active cluster name"
-	clusterStandby := "some random standby cluster name"
-	configVersion := int64(0)
-	failoverVersion := int64(59)
-	clusters := []*persistence.ClusterReplicationConfig{
-		&persistence.ClusterReplicationConfig{
-			ClusterName: clusterActive,
-		},
-		&persistence.ClusterReplicationConfig{
-			ClusterName: clusterStandby,
-		},
-	}
-
-	domainOperation := replicator.DomainOperationCreate
-	info := &persistence.DomainInfo{
-		ID:          id,
-		Name:        name,
-		Status:      persistence.DomainStatusRegistered,
-		Description: description,
-		OwnerEmail:  ownerEmail,
-	}
-	config := &persistence.DomainConfig{
-		Retention:  retention,
-		EmitMetric: emitMetric,
-	}
-	replicationConfig := &persistence.DomainReplicationConfig{
-		ActiveClusterName: clusterActive,
-		Clusters:          clusters,
-	}
-
-	s.kafkaProducer.On("Publish", &replicator.ReplicationTask{
-		TaskType: &taskType,
-		DomainTaskAttributes: &replicator.DomainTaskAttributes{
-			DomainOperation: &domainOperation,
-			ID:              common.StringPtr(id),
-			Info: &shared.DomainInfo{
-				Name:        common.StringPtr(name),
-				Status:      &status,
-				Description: common.StringPtr(description),
-				OwnerEmail:  common.StringPtr(ownerEmail),
-			},
-			Config: &shared.DomainConfiguration{
-				WorkflowExecutionRetentionPeriodInDays: common.Int32Ptr(retention),
-				EmitMetric:                             common.BoolPtr(emitMetric),
-			},
-			ReplicationConfig: &shared.DomainReplicationConfiguration{
-				ActiveClusterName: common.StringPtr(clusterActive),
-				Clusters:          s.domainReplicator.convertClusterReplicationConfigToThrift(clusters),
-			},
-			ConfigVersion:   common.Int64Ptr(configVersion),
-			FailoverVersion: common.Int64Ptr(failoverVersion),
-		},
-	}).Return(nil).Once()
-
-	err := s.domainReplicator.HandleTransmissionTask(s.kafkaProducer, domainOperation,
-		info, config, replicationConfig, configVersion, failoverVersion)
-	s.Nil(err)
-}
-
-func (s *domainReplicatorSuite) TestHandleTransmissionTask_UpdateDomainTask() {
-	taskType := replicator.ReplicationTaskTypeDomain
-	id := uuid.New()
-	name := "some random domain test name"
-	status := shared.DomainStatusDeprecated
-	description := "some random test description"
-	ownerEmail := "some random test owner"
-	retention := int32(10)
-	emitMetric := true
-	clusterActive := "some random active cluster name"
-	clusterStandby := "some random standby cluster name"
-	configVersion := int64(0)
-	failoverVersion := int64(59)
-	clusters := []*persistence.ClusterReplicationConfig{
-		&persistence.ClusterReplicationConfig{
-			ClusterName: clusterActive,
-		},
-		&persistence.ClusterReplicationConfig{
-			ClusterName: clusterStandby,
-		},
-	}
-
-	domainOperation := replicator.DomainOperationUpdate
-	info := &persistence.DomainInfo{
-		ID:          id,
-		Name:        name,
-		Status:      persistence.DomainStatusDeprecated,
-		Description: description,
-		OwnerEmail:  ownerEmail,
-	}
-	config := &persistence.DomainConfig{
-		Retention:  retention,
-		EmitMetric: emitMetric,
-	}
-	replicationConfig := &persistence.DomainReplicationConfig{
-		ActiveClusterName: clusterActive,
-		Clusters:          clusters,
-	}
-
-	s.kafkaProducer.On("Publish", &replicator.ReplicationTask{
-		TaskType: &taskType,
-		DomainTaskAttributes: &replicator.DomainTaskAttributes{
-			DomainOperation: &domainOperation,
-			ID:              common.StringPtr(id),
-			Info: &shared.DomainInfo{
-				Name:        common.StringPtr(name),
-				Status:      &status,
-				Description: common.StringPtr(description),
-				OwnerEmail:  common.StringPtr(ownerEmail),
-			},
-			Config: &shared.DomainConfiguration{
-				WorkflowExecutionRetentionPeriodInDays: common.Int32Ptr(retention),
-				EmitMetric:                             common.BoolPtr(emitMetric),
-			},
-			ReplicationConfig: &shared.DomainReplicationConfiguration{
-				ActiveClusterName: common.StringPtr(clusterActive),
-				Clusters:          s.domainReplicator.convertClusterReplicationConfigToThrift(clusters),
-			},
-			ConfigVersion:   common.Int64Ptr(configVersion),
-			FailoverVersion: common.Int64Ptr(failoverVersion),
-		},
-	}).Return(nil).Once()
-
-	err := s.domainReplicator.HandleTransmissionTask(s.kafkaProducer, domainOperation,
-		info, config, replicationConfig, configVersion, failoverVersion)
-	s.Nil(err)
 }
 
 func (s *domainReplicatorSuite) TestHandleReceivingTask_RegisterDomainTask() {
